raftstore: increase batch for raftlog-gc-worker (#11405)
* increase batch


* do not change seek


* fix fmt


* fix test


* fix number overflow


* increase batch


* revert time


* set last_compacted_idx for split region


* log large range gc task


* address comment


* close #11404

+++ components/raftstore/src/store/worker/raftlog_gc.rs
@@ -6,7 +6,7 @@ use std::sync::mpsc::Sender;
 
 use thiserror::Error;
 
-use engine_traits::{Engines, KvEngine, RaftEngine};
+use engine_traits::{Engines, KvEngine, RaftEngine, RaftLogGCTask};
 use file_system::{IOType, WithIOType};
 use tikv_util::time::{Duration, Instant};
 use tikv_util::worker::{Runnable, RunnableWithTimer};
@@ -14,6 +14,9 @@ use tikv_util::{box_try, debug, error, warn};
 
 use crate::store::worker::metrics::*;
 use crate::store::{CasualMessage, CasualRouter};
+
+const MAX_GC_REGION_BATCH: usize = 512;
+const MAX_REGION_NORMAL_GC_LOG_NUBER: u64 = 10240;
 
 pub enum Task {
     Gc {
@@ -81,13 +84,8 @@ impl<EK: KvEngine, ER: RaftEngine, R: CasualRouter<EK>> Runner<EK, ER, R> {
     }
 
     /// Does the GC job and returns the count of logs collected.
-    fn gc_raft_log(
-        &mut self,
-        region_id: u64,
-        start_idx: u64,
-        end_idx: u64,
-    ) -> Result<usize, Error> {
-        let deleted = box_try!(self.engines.raft.gc(region_id, start_idx, end_idx));
+    fn gc_raft_log(&mut self, regions: Vec<RaftLogGCTask>) -> Result<usize, Error> {
+        let deleted = box_try!(self.engines.raft.batch_gc(regions));
         Ok(deleted)
     }
 
@@ -105,8 +103,9 @@ impl<EK: KvEngine, ER: RaftEngine, R: CasualRouter<EK>> Runner<EK, ER, R> {
         });
         RAFT_LOG_GC_KV_SYNC_DURATION_HISTOGRAM.observe(start.saturating_elapsed_secs());
         let tasks = std::mem::take(&mut self.tasks);
+        let mut groups = Vec::with_capacity(tasks.len());
+        let mut need_purge = false;
         for t in tasks {
-            let start = Instant::now();
             match t {
                 Task::Gc {
                     region_id,
@@ -116,36 +115,51 @@ impl<EK: KvEngine, ER: RaftEngine, R: CasualRouter<EK>> Runner<EK, ER, R> {
                     debug!("gc raft log"; "region_id" => region_id, "end_index" => end_idx);
                     if start_idx == 0 {
                         RAFT_LOG_GC_SEEK_OPERATIONS.inc();
+                    } else if end_idx > start_idx + MAX_REGION_NORMAL_GC_LOG_NUBER {
+                        warn!("gc raft log with a large range"; "region_id" => region_id,
+                            "start_index" => start_idx,
+                            "end_index" => end_idx);
                     }
-                    match self.gc_raft_log(region_id, start_idx, end_idx) {
-                        Err(e) => {
-                            error!("failed to gc"; "region_id" => region_id, "err" => %e);
-                            self.report_collected(0);
-                            RAFT_LOG_GC_FAILED.inc();
-                        }
-                        Ok(n) => {
-                            debug!("gc log entries"; "region_id" => region_id, "entry_count" => n);
-                            self.report_collected(n);
-                            RAFT_LOG_GC_DELETED_KEYS_HISTOGRAM.observe(n as f64);
-                        }
-                    }
-                    RAFT_LOG_GC_WRITE_DURATION_HISTOGRAM.observe(start.saturating_elapsed_secs());
+                    groups.push(RaftLogGCTask {
+                        raft_group_id: region_id,
+                        from: start_idx,
+                        to: end_idx,
+                    });
                 }
                 Task::Purge => {
-                    let regions = match self.engines.raft.purge_expired_files() {
-                        Ok(regions) => regions,
-                        Err(e) => {
-                            warn!("purge expired files"; "err" => %e);
-                            return;
-                        }
-                    };
-                    for region_id in regions {
-                        let _ = self.ch.send(region_id, CasualMessage::ForceCompactRaftLogs);
-                    }
-                    RAFT_LOG_GC_PURGE_DURATION_HISTOGRAM.observe(start.saturating_elapsed_secs());
+                    need_purge = true;
                 }
             }
         }
+        let start = Instant::now();
+        match self.gc_raft_log(groups) {
+            Err(e) => {
+                error!("failed to gc"; "err" => %e);
+                self.report_collected(0);
+                RAFT_LOG_GC_FAILED.inc();
+            }
+            Ok(n) => {
+                debug!("gc log entries";  "entry_count" => n);
+                self.report_collected(n);
+                RAFT_LOG_GC_DELETED_KEYS_HISTOGRAM.observe(n as f64);
+            }
+        }
+        RAFT_LOG_GC_WRITE_DURATION_HISTOGRAM.observe(start.saturating_elapsed_secs());
+        if !need_purge {
+            return;
+        }
+        let start = Instant::now();
+        let regions = match self.engines.raft.purge_expired_files() {
+            Ok(regions) => regions,
+            Err(e) => {
+                warn!("purge expired files"; "err" => %e);
+                return;
+            }
+        };
+        for region_id in regions {
+            let _ = self.ch.send(region_id, CasualMessage::ForceCompactRaftLogs);
+        }
+        RAFT_LOG_GC_PURGE_DURATION_HISTOGRAM.observe(start.saturating_elapsed_secs());
     }
 }
 
@@ -160,6 +174,9 @@ where
     fn run(&mut self, task: Task) {
         let _io_type_guard = WithIOType::new(IOType::ForegroundWrite);
         self.tasks.push(task);
+        if self.tasks.len() > MAX_GC_REGION_BATCH {
+            self.flush();
+        }
     }
 
     fn shutdown(&mut self) {
